Devel

+++ protocols/TablesToXlsx.sh
@@ -1,4 +1,4 @@
-#MOLGENIS walltime=23:59:00 mem=26gb ppn=1 nodes=1 
+#MOLGENIS walltime=23:59:00 mem=26gb ppn=2 nodes=1 
 
 
 #string project
@@ -8,9 +8,10 @@
 #string stage
 #string tableToXlsxMod
 
-#list indelMnpTable,indelMnpRawTable,snvTable,snvRawTable
+#list indelMnpTable,indelMnpRawTable,indelMnpMinTable,indelMnpMinRawTable,snvTable,snvRawTable,snvMinTable,snvMinRawTable,svTable,svRawTable,svMinTable,svMinRawTable
 #string snvDescrTable
 #string indelMnpDescrTable
+#string svDescrTable
 
 #outXlsx
 #string xlsxDir
@@ -19,28 +20,38 @@ ${stage} ${tableToXlsxMod}
 
 outXlsx=""
 
-for inTable in "${indelMnpTable[@]}" "${indelMnpRawTable[@]}" "${snvTable[@]}" "${snvRawTable[@]}" "${snvDescrTable}" "${indelMnpDescrTable}"; do 
+#dry run???
+for inTable in "${indelMnpTable[@]}" "${indelMnpRawTable[@]}" "${indelMnpMinTable[@]}" "${indelMnpMinRawTable[@]}"\
+               "${snvTable[@]}" "${snvRawTable[@]}" "${snvMinTable[@]}" "${snvMinRawTable[@]}" \
+               "${svTable[@]}" "${svRawTable[@]}" "${svMinTable[@]}" "${svMinRawTable[@]}" \
+               "${snvDescrTable}" "${indelMnpDescrTable}"  "${svDescrTable}"; do
 	xlsx=$(echo $inTable| perl -wpe 's/.txt$|.tsv$|.csv$|.table$/.xlsx/g')
 	outXlsx="$outXlsx ${xlsxDir}/"$(basename $xlsx)
 done
 
 echo "output in the following files '$outXlsx'"
 
+#if the last file is present assume all present
 alloutputsexist \
-"$outXlsx" 
+"$outXlsx"
 
 set -x
 set -e
 
 mkdir -p ${xlsxDir}
 
-for inTable in "${indelMnpTable}" "${indelMnpRawTable}" "${snvTable}" "${snvRawTable}" "${snvDescrTable}" "${indelMnpDescrTable}"; do
-	echo "getFile file='$inTable'"
-	getFile $inTable
-	#just guess the output...
-	perl $EBROOTTABLETOXLSX/tableToXlsx.pl \\t $inTable
-	xlsx=$(echo $inTable| perl -wpe 's/.txt$|.tsv$|.csv$|.table$/.xlsx/g')
-	mv -v $xlsx ${xlsxDir}
-	putFile ${xlsxDir}/$(basename $xlsx)
-done
+for inTable in "${indelMnpTable[@]}" "${indelMnpRawTable[@]}" "${indelMnpMinTable[@]}" "${indelMnpMinRawTable[@]}"\
+               "${snvTable[@]}" "${snvRawTable[@]}" "${snvMinTable[@]}" "${snvMinRawTable[@]}" \
+               "${svTable[@]}" "${svRawTable[@]}" "${svMinTable[@]}" "${svMinRawTable[@]}" \
+               "${snvDescrTable}" "${indelMnpDescrTable}"  "${svDescrTable}"; do
 
+	if [ -f $inTable ]; then
+		echo "getFile file='$inTable'"
+		getFile $inTable
+		#just guess the output...
+		perl $EBROOTTABLETOXLSX/tableToXlsx.pl \\t $inTable
+		xlsx=$(echo $inTable| perl -wpe 's/.txt$|.tsv$|.csv$|.table$/.xlsx/g')
+		mv -v $xlsx ${xlsxDir}
+		putFile ${xlsxDir}/$(basename $xlsx)
+	fi
+done
